buildctl: add sample cli test
