+++ examples/buildkit0/buildkit.go
@@ -1,4 +1,3 @@
-// +build ignore
 package main
 
 import (
